Changed it to the recipes are individual widgets
